cli: fix stop/destroy logic
There were two issues with the destroy command: 1) it wasn't properly
associating pack names and deployment names and falsely returning a
message that no jobs were associated with a particular pack and 2) it
was ignoring var overrides when passed and greedily deleting all jobs
associated with a pack instead of the specified job name when passed as
a var override.

It also adds a stop command to maintain parity with nomad, and keeps
destroy, but as an alias for stop --purge.

Resolves #30 and #38