cli: fix meta diff in plan command
We weren't adding the job meta in the plan command, which meant that
it showed the metadata getting deleted instead of staying the
same/getting mutated. This incorporates the metadata in the plan diff
and also deletes the timestamp metadata since it wasn't required and
including it prevented runs from being idempotent.

Resolves #59