pack: perform more validation on pack before processing.
When a pack has been loaded it can go through validation to ensure
it meets basic requirement. Until now that only included metadata
validation. In situations where the root variable file was missing,
the processor would panic as the file is nil. This is now checked
in the pack validator, fixing this, and also creating a func to
continue to expand as needed.