ParsedTemplate struct
Adds a struct ParsedTemplate which holds the parse of the original job and
the canonicalized version. This allows for us to handle `namespace` and `region`,
which are defaulted in canonicalization, but require that we know the original
values when computing whether or not to take the configuration in the client.