docs: remove author and url fields from writing packs
Originally `app.author` was intended to be the actual author of the
app, but many pack maintainers put their own names here. We have the
GitHub handles of the actual maintainers, so remove this field from
the guide.

Likewise, the `pack.url` field doesn't match the actual registry URL
if someone forks a registry. Remove that field as well.