registry: store refs and source URLs of cloned repositories (#318)
This PR adds support for storing refs of cloned registry repositories as well as their full URLs. This way we avoid namespace clashes with forked repos and we'll be able to support "stale registry" detection in the future. It also means that pack no longer relies on url metadata property, which can now be safely removed.