Gigantic kitchen sink commit
This is a mess because I'd started work, then realized that
half of what made the code so inscrutable was all of the
string-keyed maps. So part of this PR is to create a PackID
and VariableID type that we can use as easier/safer map
keys.

This also gets us parsing the varfiles properly and can handle a
`-var` command line flag for a nested value